port dtnecho example to scala
This example is fully synchronous, which means it will probably break horrible in case it ever receives any messages in an unexpected non sequential order. However, the same issue seems to be present for the other examples.
Any form of communication error will cause exceptions which are not handled, which also seems fine for this example.